add wrapper script to run build.sh inside docker
It took me quite a while to figure out that 'discount' was the
markdown tool used in the build process. I've added documentation to
build.sh noting what its dependencies are.

I've also provided a wrapper script that provisions a docker container
with the correct dependencies and runs the build script. Hopefully this
will make it easier for others to submit corrections to the text and
rebuild the html files.